Made too many changes to list in the initial development of theme support. Nearing completion of an initial testing version.

+++ test/test.go
@@ -33,7 +33,7 @@ type inputJson struct {
 
 	SegwitFields [] string
 	WitnessScriptFields [] string
-	TapScriptIndex int
+	TapScriptIndex int64
 	TapScriptFields [] string
 
 	SpendType string
